added per index commands counter and detailed statistics for search, insert, replace, update commands
