+++ src/searchd.cpp
@@ -8656,6 +8656,8 @@ bool MakeSnippets ( CSphString sIndex, CSphVector<ExcerptQuery_t> & dQueries,
 		return false;
 	}
 
+	pServed->m_pStats->IncCmd ( SEARCHD_COMMAND_EXCERPT );
+
 	RIdx_c pLocalIndex { pServed };
 	assert ( pLocalIndex );
 
@@ -9256,20 +9258,20 @@ void HandleCommandUpdate ( ISphOutputBuffer & tOut, int iVer, InputBuffer_c & tR
 
 void BuildStatus ( VectorLike & dStatus )
 {
-	auto & g_tStats = gStats ();
+	const auto & tStats = gStats();
 	const char * OFF = "OFF";
 
-	const int64_t iQueriesDiv = Max ( g_tStats.m_iQueries.load ( std::memory_order_relaxed ), 1 );
-	const int64_t iDistQueriesDiv = Max ( g_tStats.m_iDistQueries.load ( std::memory_order_relaxed ), 1 );
+	const int64_t iQueriesDiv = Max ( tStats.m_iQueries.load ( std::memory_order_relaxed ), 1 );
+	const int64_t iDistQueriesDiv = Max ( tStats.m_iDistQueries.load ( std::memory_order_relaxed ), 1 );
 	const int64_t iDiv1000 = iQueriesDiv * 1000;
 	const int64_t iDDiv1000 = iDistQueriesDiv * 1000;
 
 	dStatus.SetColName ( "Counter" );
 
 	// FIXME? non-transactional!!!
-	dStatus.MatchTupletf ( "uptime", "%u", (DWORD) time ( nullptr )-g_tStats.m_uStarted );
-	dStatus.MatchTupletf ( "connections", "%l", g_tStats.m_iConnections.load ( std::memory_order_relaxed ) );
-	dStatus.MatchTupletf ( "maxed_out", "%l", g_tStats.m_iMaxedOut.load ( std::memory_order_relaxed ) );
+	dStatus.MatchTupletf ( "uptime", "%u", (DWORD) time ( nullptr )-tStats.m_uStarted );
+	dStatus.MatchTupletf ( "connections", "%l", tStats.m_iConnections.load ( std::memory_order_relaxed ) );
+	dStatus.MatchTupletf ( "maxed_out", "%l", tStats.m_iMaxedOut.load ( std::memory_order_relaxed ) );
 	dStatus.MatchTuplet ( "version" , g_sStatusVersion.cstr() );
 	dStatus.MatchTuplet ( "mysql_version", g_sMySQLVersion.cstr() );
 
@@ -9277,20 +9279,20 @@ void BuildStatus ( VectorLike & dStatus )
 	{
 		if ( i==SEARCHD_COMMAND_UNUSED_6 )
 			continue;
-		dStatus.MatchTupletf ( szCommand ( i ), "%l", g_tStats.m_iCommandCount[i].load ( std::memory_order_relaxed ) );
+		dStatus.MatchTupletf ( szCommand ( i ), "%l", tStats.Get ( (SearchdCommand_e)i ) );
 	}
 
-	FormatCmdStats ( dStatus, "insert_replace", SearchdStats_t::eReplace );
-	FormatCmdStats ( dStatus, "search", SearchdStats_t::eSearch );
-	FormatCmdStats ( dStatus, "update", SearchdStats_t::eUpdate );
+	const SearchdStats_t & tGlobalStats = gStats();
+	FormatCmdStats ( tGlobalStats, "insert_replace", SearchdStats_t::eReplace, dStatus );
+	FormatCmdStats ( tGlobalStats, "search", SearchdStats_t::eSearch, dStatus );
+	FormatCmdStats ( tGlobalStats, "update", SearchdStats_t::eUpdate, dStatus);
 
-	auto iConnects = g_tStats.m_iAgentConnectTFO.load ( std::memory_order_relaxed )
-			+g_tStats.m_iAgentConnect.load ( std::memory_order_relaxed );
+	auto iConnects = tStats.m_iAgentConnectTFO.load ( std::memory_order_relaxed ) + tStats.m_iAgentConnect.load ( std::memory_order_relaxed );
 	dStatus.MatchTupletf ( "agent_connect", "%l", iConnects );
-	dStatus.MatchTupletf ( "agent_tfo", "%l", g_tStats.m_iAgentConnectTFO.load ( std::memory_order_relaxed ) );
-	dStatus.MatchTupletf ( "agent_retry", "%l", g_tStats.m_iAgentRetry.load ( std::memory_order_relaxed ) );
-	dStatus.MatchTupletf ( "queries", "%l", g_tStats.m_iQueries.load ( std::memory_order_relaxed ) );
-	dStatus.MatchTupletf ( "dist_queries", "%l", g_tStats.m_iDistQueries.load ( std::memory_order_relaxed ) );
+	dStatus.MatchTupletf ( "agent_tfo", "%l", tStats.m_iAgentConnectTFO.load ( std::memory_order_relaxed ) );
+	dStatus.MatchTupletf ( "agent_retry", "%l", tStats.m_iAgentRetry.load ( std::memory_order_relaxed ) );
+	dStatus.MatchTupletf ( "queries", "%l", tStats.m_iQueries.load ( std::memory_order_relaxed ) );
+	dStatus.MatchTupletf ( "dist_queries", "%l", tStats.m_iDistQueries.load ( std::memory_order_relaxed ) );
 
 	// status of thread pool
 	dStatus.MatchTupletf ( "workers_total", "%d", GlobalWorkPool ()->WorkingThreads () );
@@ -9354,22 +9356,22 @@ void BuildStatus ( VectorLike & dStatus )
 		}
 	}
 
-	dStatus.MatchTupletf ( "query_wall", "%0.3F", g_tStats.m_iQueryTime.load ( std::memory_order_relaxed ) / 1000 );
+	dStatus.MatchTupletf ( "query_wall", "%0.3F", tStats.m_iQueryTime.load ( std::memory_order_relaxed ) / 1000 );
 
 	if ( g_bCpuStats )
-		dStatus.MatchTupletf ( "query_cpu", "%0.3F", g_tStats.m_iQueryCpuTime.load ( std::memory_order_relaxed ) / 1000 );
+		dStatus.MatchTupletf ( "query_cpu", "%0.3F", tStats.m_iQueryCpuTime.load ( std::memory_order_relaxed ) / 1000 );
 	else
 		dStatus.MatchTuplet ( "query_cpu", OFF);
 
-	dStatus.MatchTupletf ( "dist_wall", "%0.3F", g_tStats.m_iDistWallTime.load ( std::memory_order_relaxed ) / 1000 );
-	dStatus.MatchTupletf ( "dist_local", "%0.3F", g_tStats.m_iDistLocalTime.load ( std::memory_order_relaxed ) / 1000 );
-	dStatus.MatchTupletf ( "dist_wait", "%0.3F", g_tStats.m_iDistWaitTime.load ( std::memory_order_relaxed ) / 1000 );
+	dStatus.MatchTupletf ( "dist_wall", "%0.3F", tStats.m_iDistWallTime.load ( std::memory_order_relaxed ) / 1000 );
+	dStatus.MatchTupletf ( "dist_local", "%0.3F", tStats.m_iDistLocalTime.load ( std::memory_order_relaxed ) / 1000 );
+	dStatus.MatchTupletf ( "dist_wait", "%0.3F", tStats.m_iDistWaitTime.load ( std::memory_order_relaxed ) / 1000 );
 
 	if ( g_bIOStats )
 	{
-		dStatus.MatchTupletf ( "query_reads", "%l", g_tStats.m_iDiskReads.load ( std::memory_order_relaxed ) );
-		dStatus.MatchTupletf ( "query_readkb", "%l", g_tStats.m_iDiskReadBytes.load ( std::memory_order_relaxed )/ 1024 );
-		dStatus.MatchTupletf ( "query_readtime", "%l", g_tStats.m_iDiskReadTime.load ( std::memory_order_relaxed ) );
+		dStatus.MatchTupletf ( "query_reads", "%l", tStats.m_iDiskReads.load ( std::memory_order_relaxed ) );
+		dStatus.MatchTupletf ( "query_readkb", "%l", tStats.m_iDiskReadBytes.load ( std::memory_order_relaxed )/ 1024 );
+		dStatus.MatchTupletf ( "query_readtime", "%l", tStats.m_iDiskReadTime.load ( std::memory_order_relaxed ) );
 	} else
 	{
 		dStatus.MatchTuplet ( "query_reads", OFF );
@@ -9377,29 +9379,28 @@ void BuildStatus ( VectorLike & dStatus )
 		dStatus.MatchTuplet ( "query_readtime", OFF );
 	}
 
-	if ( g_tStats.m_iPredictedTime.load ( std::memory_order_relaxed )
-		|| g_tStats.m_iAgentPredictedTime.load ( std::memory_order_relaxed ) )
+	if ( tStats.m_iPredictedTime.load ( std::memory_order_relaxed ) || tStats.m_iAgentPredictedTime.load ( std::memory_order_relaxed ) )
 	{
-		dStatus.MatchTupletf ( "predicted_time", "%l", g_tStats.m_iPredictedTime.load ( std::memory_order_relaxed ) );
-		dStatus.MatchTupletf ( "dist_predicted_time", "%l", g_tStats.m_iAgentPredictedTime.load ( std::memory_order_relaxed ) );
+		dStatus.MatchTupletf ( "predicted_time", "%l", tStats.m_iPredictedTime.load ( std::memory_order_relaxed ) );
+		dStatus.MatchTupletf ( "dist_predicted_time", "%l", tStats.m_iAgentPredictedTime.load ( std::memory_order_relaxed ) );
 	}
 
-	dStatus.MatchTupletf ( "avg_query_wall", "%0.3F", g_tStats.m_iQueryTime.load ( std::memory_order_relaxed ) / iDiv1000 );
+	dStatus.MatchTupletf ( "avg_query_wall", "%0.3F", tStats.m_iQueryTime.load ( std::memory_order_relaxed ) / iDiv1000 );
 
 	if ( g_bCpuStats )
-		dStatus.MatchTupletf ( "avg_query_cpu", "%0.3F", g_tStats.m_iQueryCpuTime.load ( std::memory_order_relaxed ) / iDiv1000 );
+		dStatus.MatchTupletf ( "avg_query_cpu", "%0.3F", tStats.m_iQueryCpuTime.load ( std::memory_order_relaxed ) / iDiv1000 );
 	else
 		dStatus.MatchTuplet ( "avg_query_cpu", OFF );
 
-	dStatus.MatchTupletf ( "avg_dist_wall", "%0.3F", g_tStats.m_iDistWallTime.load ( std::memory_order_relaxed ) / iDDiv1000 );
-	dStatus.MatchTupletf ( "avg_dist_local", "%0.3F", g_tStats.m_iDistLocalTime.load ( std::memory_order_relaxed ) / iDDiv1000 );
-	dStatus.MatchTupletf ( "avg_dist_wait", "%0.3F", g_tStats.m_iDistWaitTime.load ( std::memory_order_relaxed ) / iDDiv1000 );
+	dStatus.MatchTupletf ( "avg_dist_wall", "%0.3F", tStats.m_iDistWallTime.load ( std::memory_order_relaxed ) / iDDiv1000 );
+	dStatus.MatchTupletf ( "avg_dist_local", "%0.3F", tStats.m_iDistLocalTime.load ( std::memory_order_relaxed ) / iDDiv1000 );
+	dStatus.MatchTupletf ( "avg_dist_wait", "%0.3F", tStats.m_iDistWaitTime.load ( std::memory_order_relaxed ) / iDDiv1000 );
 
 	if ( g_bIOStats )
 	{
-		dStatus.MatchTupletf ( "avg_query_reads", "%0.1F", g_tStats.m_iDiskReads.load ( std::memory_order_relaxed ) * 10 / iQueriesDiv );
-		dStatus.MatchTupletf ( "avg_query_readkb", "%0.1F", g_tStats.m_iDiskReadBytes.load ( std::memory_order_relaxed ) * 10 / (iQueriesDiv*1024) );
-		dStatus.MatchTupletf ( "avg_query_readtime", "%0.3F", g_tStats.m_iDiskReadTime.load ( std::memory_order_relaxed ) / iDiv1000 );
+		dStatus.MatchTupletf ( "avg_query_reads", "%0.1F", tStats.m_iDiskReads.load ( std::memory_order_relaxed ) * 10 / iQueriesDiv );
+		dStatus.MatchTupletf ( "avg_query_readkb", "%0.1F", tStats.m_iDiskReadBytes.load ( std::memory_order_relaxed ) * 10 / (iQueriesDiv*1024) );
+		dStatus.MatchTupletf ( "avg_query_readtime", "%0.3F", tStats.m_iDiskReadTime.load ( std::memory_order_relaxed ) / iDiv1000 );
 	} else
 	{
 		dStatus.MatchTuplet ( "avg_query_reads", OFF );
@@ -10399,6 +10400,8 @@ static void PQLocalMatch ( const BlobVec_t & dDocs, const CSphString & sIndex, c
 	if ( pServed->m_eType!=IndexType_e::PERCOLATE )
 		return sMsg.Err ( "table '%s' is not percolate", sIndex.cstr () );
 
+	pServed->m_pStats->IncCmd ( SEARCHD_COMMAND_CALLPQ );
+
 	RIdx_T<PercolateIndex_i*> pIndex { pServed };
 	RtAccum_t * pAccum = tAcc.GetAcc ( pIndex, sError );
 	sMsg.Err ( sError );
@@ -11385,13 +11388,16 @@ static bool CleanupAcc ( bool bMissed, RtAccum_t * pAccum, CSphString & sError )
 	return false;
 }
 
-static bool CheckAccIndex ( CSphSessionAccum & tSession, CSphString & sError )
+static bool CheckAccIndex ( CSphSessionAccum & tSession, CSphString & sError, bool bIncServedCmd )
 {
 	RtAccum_t * pAccum = tSession.GetAcc();
 	assert ( pAccum );
 	auto pServed = GetServed ( pAccum->GetIndexName() );
 	if ( !pServed )
 		return CleanupAcc ( true, pAccum, sError );
+
+	if ( bIncServedCmd )
+		pServed->m_pStats->IncCmd ( SEARCHD_COMMAND_COMMIT );
 
 	if ( pAccum->GetIndexId()!=RIdx_T<RtIndex_i*>( pServed )->GetIndexId() )
 		return CleanupAcc ( false, pAccum, sError );
@@ -11408,7 +11414,7 @@ void sphHandleMysqlBegin ( StmtErrorReporter_i& tOut, Str_t sQuery )
 	MEMORY ( MEM_SQL_BEGIN );
 	if ( tAcc.GetIndex() )
 	{
-		if ( !CheckAccIndex ( tAcc, sError ) )
+		if ( !CheckAccIndex ( tAcc, sError, false ) )
 			return tOut.Error ( "%s", sError.cstr() );
 
 		if ( !HandleCmdReplicate ( *tAcc.GetAcc() ) )
@@ -11438,7 +11444,7 @@ void sphHandleMysqlCommitRollback ( StmtErrorReporter_i& tOut, Str_t sQuery, boo
 		RtAccum_t * pAccum = tAcc.GetAcc();
 		tCrashQuery.m_dIndex = FromStr ( pAccum->GetIndexName() );
 
-		if ( !CheckAccIndex ( tAcc, sError ) )
+		if ( !CheckAccIndex ( tAcc, sError, true ) )
 			return tOut.Error ( "%s", sError.cstr() );
 
 		if ( bCommit )
@@ -11500,7 +11506,8 @@ void sphHandleMysqlInsert ( StmtErrorReporter_i & tOut, const SqlStmt_t & tStmt 
 
 	// index lock after replication takes place
 	CommitAcc ( tStmt, pServed, tOut );
-	StatCountCommandDetails ( SearchdStats_t::eReplace, tStmt.m_iRowsAffected, tmStart );
+	pServed->m_pStats->AddWriteStat ( SearchdStats_t::eReplace, bReplace, tStmt.m_iRowsAffected, tmStart );
+	gStats().AddDetailed ( SearchdStats_t::eReplace, tStmt.m_iRowsAffected, tmStart );
 }
 
 // when index name came as `cluster:name`, it came to the name, and should be splitted to cluster and index name
@@ -11895,8 +11902,10 @@ bool DoGetKeywords ( const CSphString & sIndex, const CSphString & sQuery, const
 	bool bOk = false;
 	// just local plain or template index
 	if ( pLocal )
+	{
+		pLocal->m_pStats->IncCmd ( SEARCHD_COMMAND_KEYWORDS );
 		bOk = RIdx_c(pLocal)->GetKeywords ( dKeywords, sQuery.cstr(), tSettings, &sError );
-	else
+	} else
 	{
 		// FIXME!!! g_iDistThreads thread pool for locals.
 		// locals
@@ -11910,6 +11919,7 @@ bool DoGetKeywords ( const CSphString & sIndex, const CSphString & sQuery, const
 				tFailureLog.Submit ( sLocal.cstr(), sIndex.cstr(), "missed table" );
 				continue;
 			}
+			pServed->m_pStats->IncCmd ( SEARCHD_COMMAND_KEYWORDS );
 
 			dKeywordsLocal.Resize(0);
 			if ( RIdx_c ( pServed )->GetKeywords ( dKeywordsLocal, sQuery.cstr(), tSettings, &sError ) )
@@ -12287,6 +12297,9 @@ static void SuggestSendResult ( const SuggestArgs_t & tArgs, SuggestResultSet_t 
 
 static bool SuggestLocalIndexGet ( const cServedIndexRefPtr_c & pServed, const SuggestArgs_t & tArgs, const char * sWord, SuggestResult_t & tRes, CSphString & sError )
 {
+	assert ( pServed );
+	pServed->m_pStats->IncCmd ( SEARCHD_COMMAND_SUGGEST );
+
 	RIdx_c pIdx { pServed };
 	if ( !pIdx->GetSettings().m_iMinInfixLen || !pIdx->GetDictionary()->GetSettings().m_bWordDict )
 	{
@@ -13879,7 +13892,13 @@ void sphHandleMysqlUpdate ( StmtErrorReporter_i & tOut, const SqlStmt_t & tStmt,
 		auto pLocked = GetServed ( sReqIndex );
 		if ( pLocked )
 		{
-			DoExtendedUpdate ( tStmt, sReqIndex, nullptr, bBlobUpdate, iSuccesses, iUpdated, dFails, sWarning, pLocked );
+			int64_t tmStartIdx = sphMicroTimer();
+			int iUpdatedIdx = 0;
+
+			DoExtendedUpdate ( tStmt, sReqIndex, nullptr, bBlobUpdate, iSuccesses, iUpdatedIdx, dFails, sWarning, pLocked );
+
+			iUpdated += iUpdatedIdx;
+			pLocked->m_pStats->AddWriteStat ( SearchdStats_t::eUpdate, false, iUpdatedIdx, tmStartIdx );
 
 		} else if ( dDistributed[iIdx] )
 		{
@@ -13889,8 +13908,15 @@ void sphHandleMysqlUpdate ( StmtErrorReporter_i & tOut, const SqlStmt_t & tStmt,
 			ARRAY_FOREACH ( i, dLocal )
 			{
 				const char * sLocal = dLocal[i].cstr();
+				int64_t tmStartIdx = sphMicroTimer();
+				int iUpdatedIdx = 0;
+
 				auto pServed = GetServed ( sLocal );
-				DoExtendedUpdate ( tStmt, sLocal, sReqIndex, bBlobUpdate, iSuccesses, iUpdated, dFails, sWarning, pServed );
+				DoExtendedUpdate ( tStmt, sLocal, sReqIndex, bBlobUpdate, iSuccesses, iUpdatedIdx, dFails, sWarning, pServed );
+
+				iUpdated += iUpdatedIdx;
+				if ( pServed )
+					pServed->m_pStats->AddWriteStat ( SearchdStats_t::eUpdate, false, iUpdatedIdx, tmStartIdx );
 			}
 		}
 
@@ -13912,7 +13938,7 @@ void sphHandleMysqlUpdate ( StmtErrorReporter_i & tOut, const SqlStmt_t & tStmt,
 	StringBuilder_c sReport;
 	dFails.BuildReport ( sReport );
 
-	StatCountCommandDetails ( SearchdStats_t::eUpdate, iUpdated, tmStart );
+	gStats().AddDetailed ( SearchdStats_t::eUpdate, iUpdated, tmStart );
 	if ( !iSuccesses )
 	{
 		tOut.Error ( "%s", sReport.cstr() );
@@ -14484,6 +14510,8 @@ static int LocalIndexDoDeleteDocuments ( const CSphString & sName, const char * 
 	if ( !ServedDesc_t::IsMutable ( pServed ) )
 		return err ( "table not available, or does not support DELETE" );
 
+	pServed->m_pStats->IncCmd ( SEARCHD_COMMAND_DELETE );
+
 	GlobalCrashQueryGetRef().m_dIndex = FromStr ( sName );
 	if ( !ValidateClusterStatement ( sName, *pServed, sCluster, IsHttpStmt ( tStmt ) ) )
 		return err ( TlsMsg::szError() );
@@ -14799,7 +14827,7 @@ void HandleMysqlMultiStmt ( const CSphVector<SqlStmt_t> & dStmt, CSphQueryResult
 				bMoreResultsFollow = false;
 
 			auto uMatches = SendMysqlSelectResult ( dRows, tRes, bMoreResultsFollow, false, nullptr, ( tSess.IsProfile() ? &tProfile : nullptr ) );
-			StatCountCommandDetails ( SearchdStats_t::eSearch, uMatches, tmStart );
+			gStats().AddDetailed ( SearchdStats_t::eSearch, uMatches, tmStart );
 			break;
 		}
 		case STMT_SHOW_WARNINGS:
@@ -16035,19 +16063,29 @@ static void AddFoundRowsStatsToOutput ( VectorLike & dStatus, const char * szPre
 }
 
 
-static void AddIndexQueryStats ( VectorLike & dStatus, const ServedStats_c& tStats )
+void ServedStats_c::GetIndexQueryStats ( VectorLike & dStatus ) const
 {
 	QueryStats_t tQueryTimeStats, tRowsFoundStats;
-	tStats.CalculateQueryStats ( tRowsFoundStats, tQueryTimeStats );
+	CalculateQueryStats ( tRowsFoundStats, tQueryTimeStats );
 	AddQueryTimeStatsToOutput ( dStatus, "query_time", tQueryTimeStats );
 
 #ifndef NDEBUG
 	QueryStats_t tExactQueryTimeStats, tExactRowsFoundStats;
-	tStats.CalculateQueryStatsExact ( tExactQueryTimeStats, tExactRowsFoundStats );
+	CalculateQueryStatsExact ( tExactQueryTimeStats, tExactRowsFoundStats );
 	AddQueryTimeStatsToOutput ( dStatus, "exact_query_time", tQueryTimeStats );
 #endif
 
 	AddFoundRowsStatsToOutput ( dStatus, "found_rows", tRowsFoundStats );
+
+	// command stats
+	SearchdCommand_e dCommands[] = { SEARCHD_COMMAND_SEARCH, SEARCHD_COMMAND_EXCERPT, SEARCHD_COMMAND_UPDATE, SEARCHD_COMMAND_KEYWORDS, SEARCHD_COMMAND_STATUS, SEARCHD_COMMAND_DELETE, SEARCHD_COMMAND_INSERT, SEARCHD_COMMAND_REPLACE, SEARCHD_COMMAND_COMMIT, SEARCHD_COMMAND_SUGGEST, SEARCHD_COMMAND_CALLPQ, SEARCHD_COMMAND_GETFIELD };
+	for ( auto eCmd : dCommands )
+		dStatus.MatchTupletf ( szCommand ( eCmd ), "%l", m_tCommandsStats.Get ( eCmd ) );
+
+	// detailed stats
+	FormatCmdStats ( m_tCommandsStats, "insert_replace", SearchdStats_t::eReplace, dStatus );
+	FormatCmdStats ( m_tCommandsStats, "search", SearchdStats_t::eSearch, dStatus );
+	FormatCmdStats ( m_tCommandsStats, "update", SearchdStats_t::eUpdate, dStatus );
 }
 
 static void AddDiskIndexStatus ( VectorLike & dStatus, const CSphIndex * pIndex, bool bRt, bool bPq )
@@ -16147,7 +16185,7 @@ static void AddPlainIndexStatus ( RowBuffer_i & tOut, const cServedIndexRefPtr_c
 	if ( pServed->m_eType != IndexType_e::TEMPLATE )
 	{
 		AddDiskIndexStatus ( dStatus, pIndex, pServed->m_eType == IndexType_e::RT, pServed->m_eType == IndexType_e::PERCOLATE );
-		AddIndexQueryStats ( dStatus, tStats );
+		tStats.GetIndexQueryStats ( dStatus );
 	}
 	tOut.DataTable ( dStatus );
 }
@@ -16158,7 +16196,7 @@ static void AddDistibutedIndexStatus ( RowBuffer_i & tOut, const cDistributedInd
 	assert ( pIndex );
 	VectorLike dStatus ( sPattern );
 	dStatus.MatchTuplet( "table_type", "distributed" );
-	AddIndexQueryStats ( dStatus, pIndex->m_tStats );
+	pIndex->m_tStats.GetIndexQueryStats ( dStatus );
 	tOut.DataTable ( dStatus );
 }
 
@@ -16173,6 +16211,8 @@ void HandleMysqlShowIndexStatus ( RowBuffer_i& tOut, const SqlStmt_t& tStmt )
 
 	if ( pServed )
 	{
+		pServed->m_pStats->IncCmd ( SEARCHD_COMMAND_STATUS );
+
 		if ( iChunk >= 0 && pServed->m_eType == IndexType_e::RT )
 		{
 			RIdx_T<const RtIndex_i*> ( pServed )->ProcessDiskChunk ( iChunk, [&tOut, &tStmt] ( const CSphIndex* pIndex ) {
@@ -16194,9 +16234,13 @@ void HandleMysqlShowIndexStatus ( RowBuffer_i& tOut, const SqlStmt_t& tStmt )
 	auto pIndex = GetDistr ( tStmt.m_sIndex );
 
 	if ( pIndex )
+	{
+		pIndex->m_tStats.IncCmd ( SEARCHD_COMMAND_STATUS );
 		AddDistibutedIndexStatus ( tOut, pIndex, tStmt.m_sIndex, tStmt.m_sStringParam );
-	else
+	} else
+	{
 		tOut.Error ( "SHOW TABLE STATUS requires an existing table" );
+	}
 }
 
 static bool AddFederatedIndexStatusHeader ( RowBuffer_i& tOut )
@@ -17617,7 +17661,7 @@ bool ClientSession_c::Execute ( Str_t sQuery, RowBuffer_i & tOut )
 				m_sError = "";
 				AggrResult_t & tLast = tHandler.m_dAggrResults.Last();
 				auto uMatches = SendMysqlSelectResult ( tOut, tLast, false, m_bFederatedUser, &m_sFederatedQuery, ( tSess.IsProfile() ? &m_tProfile : nullptr ) );
-				StatCountCommandDetails ( SearchdStats_t::eSearch, uMatches, tmStart );
+				gStats().AddDetailed ( SearchdStats_t::eSearch, uMatches, tmStart );
 			}
 
 			// save meta for SHOW META (profile is saved elsewhere)
@@ -18149,8 +18193,7 @@ void HandleCommandJson ( ISphOutputBuffer & tOut, WORD uVer, InputBuffer_c & tRe
 
 void StatCountCommand ( SearchdCommand_e eCmd )
 {
-	if ( eCmd<SEARCHD_COMMAND_TOTAL )
-		gStats ().m_iCommandCount[eCmd].fetch_add ( 1, std::memory_order_relaxed );
+	gStats().Inc ( eCmd );
 }
 
 // fixme! move federated stuff to another parser and remove all kind of 'fixups'
