+++ src/searchdaemon.cpp
@@ -1042,183 +1042,10 @@ void InputBuffer_c::SetMaxPacketSize( int iMaxPacketSize )
 // SERVED INDEX DESCRIPTORS STUFF
 /////////////////////////////////////////////////////////////////////////////
 
-
-class QueryStatContainer_c: public QueryStatContainer_i
-{
-public:
-	void Add( uint64_t uFoundRows, uint64_t uQueryTime, uint64_t uTimestamp ) final;
-	QueryStatRecord_t GetRecord( int iRecord ) const noexcept final;
-	int GetNumRecords() const final;
-
-	QueryStatContainer_c();
-	QueryStatContainer_c( QueryStatContainer_c&& tOther ) noexcept;
-	void Swap( QueryStatContainer_c& rhs ) noexcept;
-	QueryStatContainer_c& operator=( QueryStatContainer_c tOther ) noexcept;
-
-private:
-	CircularBuffer_T<QueryStatRecord_t> m_dRecords;
-};
-
-std::unique_ptr<QueryStatContainer_i> MakeStatsContainer ()
-{
-	return std::make_unique<QueryStatContainer_c>();
-}
-
-void QueryStatContainer_c::Add( uint64_t uFoundRows, uint64_t uQueryTime, uint64_t uTimestamp )
-{
-	if ( !m_dRecords.IsEmpty())
-	{
-		QueryStatRecord_t& tLast = m_dRecords.Last();
-		const uint64_t BUCKET_TIME_DELTA = 100000;
-		if ( uTimestamp - tLast.m_uTimestamp<=BUCKET_TIME_DELTA )
-		{
-			tLast.m_uFoundRowsMin = Min( uFoundRows, tLast.m_uFoundRowsMin );
-			tLast.m_uFoundRowsMax = Max( uFoundRows, tLast.m_uFoundRowsMax );
-			tLast.m_uFoundRowsSum += uFoundRows;
-
-			tLast.m_uQueryTimeMin = Min( uQueryTime, tLast.m_uQueryTimeMin );
-			tLast.m_uQueryTimeMax = Max( uQueryTime, tLast.m_uQueryTimeMax );
-			tLast.m_uQueryTimeSum += uQueryTime;
-
-			tLast.m_iCount++;
-
-			return;
-		}
-	}
-
-	const uint64_t MAX_TIME_DELTA = 15 * 60 * 1000000;
-	while ( !m_dRecords.IsEmpty() && ( uTimestamp - m_dRecords[0].m_uTimestamp )>MAX_TIME_DELTA )
-		m_dRecords.Pop();
-
-	QueryStatRecord_t& tRecord = m_dRecords.Push();
-	tRecord.m_uFoundRowsMin = uFoundRows;
-	tRecord.m_uFoundRowsMax = uFoundRows;
-	tRecord.m_uFoundRowsSum = uFoundRows;
-
-	tRecord.m_uQueryTimeMin = uQueryTime;
-	tRecord.m_uQueryTimeMax = uQueryTime;
-	tRecord.m_uQueryTimeSum = uQueryTime;
-
-	tRecord.m_uTimestamp = uTimestamp;
-	tRecord.m_iCount = 1;
-}
-
-QueryStatRecord_t QueryStatContainer_c::GetRecord ( int iRecord ) const noexcept
-{
-	return m_dRecords[iRecord];
-}
-
-
-int QueryStatContainer_c::GetNumRecords() const
-{
-	return m_dRecords.GetLength();
-}
-
-QueryStatContainer_c::QueryStatContainer_c() = default;
-
-QueryStatContainer_c::QueryStatContainer_c( QueryStatContainer_c&& tOther ) noexcept
-	: QueryStatContainer_c()
-{ Swap( tOther ); }
-
-void QueryStatContainer_c::Swap( QueryStatContainer_c& rhs ) noexcept
-{
-	rhs.m_dRecords.Swap( m_dRecords );
-}
-
-QueryStatContainer_c& QueryStatContainer_c::operator=( QueryStatContainer_c tOther ) noexcept
-{
-	Swap( tOther );
-	return *this;
-}
-
-//////////////////////////////////////////////////////////////////////////
-
-#ifndef NDEBUG
-
-class QueryStatContainerExact_c: public QueryStatContainer_i
-{
-public:
-	void Add( uint64_t uFoundRows, uint64_t uQueryTime, uint64_t uTimestamp ) final;
-	QueryStatRecord_t GetRecord( int iRecord ) const noexcept final;
-	int GetNumRecords() const final;
-
-	QueryStatContainerExact_c();
-	QueryStatContainerExact_c( QueryStatContainerExact_c&& tOther ) noexcept;
-	void Swap( QueryStatContainerExact_c& rhs ) noexcept;
-	QueryStatContainerExact_c& operator=( QueryStatContainerExact_c tOther ) noexcept;
-
-private:
-	struct QueryStatRecordExact_t
-	{
-		uint64_t m_uQueryTime;
-		uint64_t m_uFoundRows;
-		uint64_t m_uTimestamp;
-	};
-
-	CircularBuffer_T<QueryStatRecordExact_t> m_dRecords;
-};
-
-void QueryStatContainerExact_c::Add( uint64_t uFoundRows, uint64_t uQueryTime, uint64_t uTimestamp )
-{
-	const uint64_t MAX_TIME_DELTA = 15 * 60 * 1000000;
-	while ( !m_dRecords.IsEmpty() && ( uTimestamp - m_dRecords[0].m_uTimestamp )>MAX_TIME_DELTA )
-		m_dRecords.Pop();
-
-	QueryStatRecordExact_t& tRecord = m_dRecords.Push();
-	tRecord.m_uFoundRows = uFoundRows;
-	tRecord.m_uQueryTime = uQueryTime;
-	tRecord.m_uTimestamp = uTimestamp;
-}
-
-
-int QueryStatContainerExact_c::GetNumRecords() const
-{
-	return m_dRecords.GetLength();
-}
-
-
-QueryStatRecord_t QueryStatContainerExact_c::GetRecord ( int iRecord ) const noexcept
-{
-	QueryStatRecord_t tRecord;
-	const QueryStatRecordExact_t& tExact = m_dRecords[iRecord];
-
-	tRecord.m_uQueryTimeMin = tExact.m_uQueryTime;
-	tRecord.m_uQueryTimeMax = tExact.m_uQueryTime;
-	tRecord.m_uQueryTimeSum = tExact.m_uQueryTime;
-	tRecord.m_uFoundRowsMin = tExact.m_uFoundRows;
-	tRecord.m_uFoundRowsMax = tExact.m_uFoundRows;
-	tRecord.m_uFoundRowsSum = tExact.m_uFoundRows;
-
-	tRecord.m_uTimestamp = tExact.m_uTimestamp;
-	tRecord.m_iCount = 1;
-	return tRecord;
-}
-
-QueryStatContainerExact_c::QueryStatContainerExact_c() = default;
-
-QueryStatContainerExact_c::QueryStatContainerExact_c( QueryStatContainerExact_c&& tOther ) noexcept
-	: QueryStatContainerExact_c()
-{ Swap( tOther ); }
-
-void QueryStatContainerExact_c::Swap( QueryStatContainerExact_c& rhs ) noexcept
-{
-	rhs.m_dRecords.Swap( m_dRecords );
-}
-
-QueryStatContainerExact_c& QueryStatContainerExact_c::operator=( QueryStatContainerExact_c tOther ) noexcept
-{
-	Swap( tOther );
-	return *this;
-}
-
-#endif
-
-
-//////////////////////////////////////////////////////////////////////////
 ServedStats_c::ServedStats_c()
-	: m_pQueryStatRecords { std::make_unique<QueryStatContainer_c>() }
+	: m_pQueryStatRecords { MakeStatsContainer() }
 #ifndef NDEBUG
-	, m_pQueryStatRecordsExact { std::make_unique<QueryStatContainerExact_c>() }
+	, m_pQueryStatRecordsExact { MakeExactStatsContainer() }
 #endif
 
 {}
@@ -1245,9 +1072,11 @@ void ServedStats_c::AddQueryStat( uint64_t uFoundRows, uint64_t uQueryTime )
 	m_uTotalQueryTimeMax = Max( uQueryTime, m_uTotalQueryTimeMax );
 	m_uTotalQueryTimeSum += uQueryTime;
 
+	m_tCommandsStats.AddDeltaDetailed ( CommandStats_t::eSearch, uFoundRows, uQueryTime );
+	m_tCommandsStats.Inc ( SEARCHD_COMMAND_SEARCH );
+
 	++m_uTotalQueries;
 }
-
 
 static const uint64_t g_dStatsIntervals[] =
 	{
@@ -1373,6 +1202,27 @@ void ServedStats_c::DoStatCalcStats( const QueryStatContainer_i* pContainer, Que
 	tQueryAllStats.m_dData[TYPE_95] = ( uint64_t ) m_tQueryTimeDigest.Percentile( 95 );
 	tQueryAllStats.m_dData[TYPE_99] = ( uint64_t ) m_tQueryTimeDigest.Percentile( 99 );
 	tQueryAllStats.m_uTotalQueries = m_uTotalQueries;
+}
+
+void ServedStats_c::AddWriteStat ( CommandStats_t::EDETAILS eCmd, bool bReplace, uint64_t uRows, uint64_t tmStart )
+{
+	m_tCommandsStats.AddDetailed ( eCmd, uRows, tmStart );
+	switch ( eCmd )
+	{
+	case CommandStats_t::eUpdate:
+		m_tCommandsStats.Inc ( SEARCHD_COMMAND_UPDATE );
+		break;
+	case CommandStats_t::eReplace:
+		m_tCommandsStats.Inc ( bReplace ? SEARCHD_COMMAND_REPLACE : SEARCHD_COMMAND_INSERT );
+		break;
+
+	default: break;
+	}
+}
+
+void ServedStats_c::IncCmd ( SearchdCommand_e eCmd )
+{
+	m_tCommandsStats.Inc ( eCmd );
 }
 
 //////////////////////////////////////////////////////////////////////////
